expander start

+++ ft_strlen.c
@@ -1,9 +1,9 @@
 
 #include "minishell.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	count;
+	int	count;
 
 	count = 0;
 	if (!s)
